% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_candidate_signature.R
\name{check_query_signature}
\alias{check_query_signature}
\title{Verify if the input data.frame of the query signature contains both the gene_id and zscore columns.}
\usage{
check_query_signature(query_signature)
}
\arguments{
\item{query_signature}{\code{data.frame} object}
}
\value{
A \code{logical} object
}
\description{
Verify if the input data.frame of the query signature contains both the gene_id and zscore columns.
}
\examples{
data(query_signature)
check_query_signature(query_signature)

}
