% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/web_request.R
\name{checkExistenceInCmapDown}
\alias{checkExistenceInCmapDown}
\title{checkExistenceInCmapDown}
\usage{
checkExistenceInCmapDown(gene_id)
}
\arguments{
\item{gene_id}{the ENTREZ id}
}
\value{
candidate signature index
}
\description{
checkExistenceInCmapDown
}
