% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/web_request.R
\name{checkExistenceInCmapUp}
\alias{checkExistenceInCmapUp}
\title{checkExistenceInCmapUp}
\usage{
checkExistenceInCmapUp(gene_id)
}
\arguments{
\item{gene_id}{the ENTREZ id}
}
\value{
candidate signature index
}
\description{
checkExistenceInCmapUp
}
\examples{
data(query_signature)
K=50
checkExistenceInCmapUp(gene_id="348")

}
