% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{gsea_plot}
\alias{gsea_plot}
\title{gsea plot}
\usage{
gsea_plot(data, query_signature, K = 50)
}
\arguments{
\item{data}{the selected connectivity analysis result \code{data.frame} to plot}
}
\value{
The ggplot object
}
\description{
gsea plot
}
