% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Connectivity_analysis.R
\name{positive_query_CMap}
\alias{positive_query_CMap}
\title{To identify candidate signatures that are positively correlated with the query signature by computing six kinds of connectivity scores}
\usage{
positive_query_CMap(signature_index)
}
\arguments{
\item{signature_index}{}
}
\value{
a \code{data.frame}
}
\description{
To identify candidate signatures that are positively correlated with the query signature by computing six kinds of connectivity scores
}
\examples{
data(query_signature2)
query_signature = query_signature2
K=50
res = positive_query_CMap(519170)

}
