% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{query_signature}
\alias{query_signature}
\title{Example query signature}
\format{
A data frame with 11,689 rows and 2 variables:
\describe{
\item{gene_id}{gene Entrez ID, character}
\item{zscore}{derived from comparing perturbation and control conditions to quantify both the direction and magnitude of transcriptional changes, numeric}
}
}
\source{
Generated internally
}
\usage{
data(query_signature)
}
\description{
An example gene signatures for demonstration.
}
\keyword{datasets}
