% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{radar_plot}
\alias{radar_plot}
\title{radar plot}
\usage{
radar_plot(data)
}
\arguments{
\item{data}{the selected connectivity analysis result \code{data.frame} to plot}
}
\value{
The ggplot object
}
\description{
radar plot
}
