% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/web_request.R
\name{retrieveCmapSignature}
\alias{retrieveCmapSignature}
\title{Retrieve CMap signature from remote server}
\usage{
retrieveCmapSignature(signature_index)
}
\arguments{
\item{signature_index}{}
}
\value{
a \code{data.frame} object
}
\description{
Retrieve CMap signature from remote server
}
