% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/web_request.R
\name{retrieveCmapSignatureMeta}
\alias{retrieveCmapSignatureMeta}
\title{Retrieve CMap signature meta information from remote server}
\usage{
retrieveCmapSignatureMeta(signature_index)
}
\arguments{
\item{signature_index}{}
}
\value{
a \code{data.frame} object
}
\description{
Retrieve CMap signature meta information from remote server
}
\examples{

K=50
retrieveCmapSignatureMeta(signature_index=519170)

}
