% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/web_request.R
\name{retrievePermutationResult}
\alias{retrievePermutationResult}
\title{retrieve the Permutation result from remote server}
\usage{
retrievePermutationResult(signature_index, method)
}
\arguments{
\item{signature_index}{the index of CMap signature}

\item{method}{XSum or CSS}
}
\value{
a \code{data.frame}
}
\description{
retrieve the Permutation result from remote server
}
