% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectivity_analysis.R
\name{run_positive_CMap}
\alias{run_positive_CMap}
\title{run positive connectivity analysis between query signature and Extend CMap compound-induced signature}
\usage{
run_positive_CMap(query_signature, K = 50, cores = 5)
}
\arguments{
\item{query_signature}{a \code{data.frame} must include the gene_id and zscore columns.}

\item{K}{default 50}

\item{cores}{default 5}
}
\value{
a \code{data.frame} object
}
\description{
run positive connectivity analysis between query signature and Extend CMap compound-induced signature
}
