% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_candidate_signature.R
\name{screen_negative_candidate_signature}
\alias{screen_negative_candidate_signature}
\title{screening negative candidate signatures}
\usage{
screen_negative_candidate_signature(query_signature, K = 50, cores = 5)
}
\arguments{
\item{query_signature}{input}

\item{K}{the top K and bottom K genes}

\item{cores}{the cores used for computation}
}
\value{
a \code{vector}
}
\description{
screening negative candidate signatures
}
\examples{

data(query_signature)
K=50
screen_negative_candidate_signature(query_signature,K=50,cores=1)

}
