% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{SYMBOLtoENTREZ}
\alias{SYMBOLtoENTREZ}
\title{Convert gene Symbol to ENTREZ ID}
\usage{
SYMBOLtoENTREZ(x)
}
\arguments{
\item{a}{vector of gene symbols}
}
\value{
a vector of ENTREZ IDs
}
\description{
Convert gene Symbol to ENTREZ ID
}
\examples{
entrezid <- SYMBOLtoENTREZ("APOE")
}
