% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WebCMap-package.R
\name{WebCMap}
\alias{WebCMap}
\title{WebCMap: An ultra-fast connectivity analysis tool using web technologies to accelerate computations based on the extended Connectivity Map (CMap).}
\description{
An ultra-fast connectivity analysis tool using web technologies to accelerate computations based on the extended Connectivity Map (CMap).
}
\author{
Hongen Kang  \email{kanghongen@gmail.com}
}
\keyword{internal}
