% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{venn_plot}
\alias{venn_plot}
\title{venn plot}
\usage{
venn_plot(data, query_signature, K = 50)
}
\arguments{
\item{data}{the selected connectivity analysis result \code{data.frame} to plot}
}
\value{
The ggplot object
}
\description{
venn plot
}
